+++ examples/azure/main.tf
@@ -1,25 +1,25 @@
 terraform {
-  required_version = "~> 0.13"
+  required_version = "~> 1.5"
   required_providers {
     azuread = {
       source  = "hashicorp/azuread"
-      version = "~> 1.0"
+      version = "~> 2.47"
     }
     azurerm = {
       source  = "hashicorp/azurerm"
-      version = "~> 2.34.0"
+      version = "~> 3.85"
     }
     mssql = {
       source  = "betr-io/mssql"
-      version = "0.1.0"
+      version = "~> 0.2"
     }
     random = {
       source  = "hashicorp/random"
-      version = "~> 3.0.0"
+      version = "~> 3.6"
     }
     time = {
       source  = "hashicorp/time"
-      version = "0.6.0"
+      version = "~> 0.10"
     }
   }
 }
@@ -78,57 +78,41 @@ locals {
   prefix = "${var.prefix}-${substr(random_string.random.result, 0, 4)}"
 }
 
-resource "random_password" "sa" {
-  length  = 32
-  special = true
-  keepers = {
-    name = "${local.prefix}-sa"
-  }
-}
-
-resource "random_password" "user" {
-  length  = 32
-  special = true
-  keepers = {
-    name = "${local.prefix}-user"
-  }
-}
-
 # An Azure AD group assigned the role 'Directory Readers'. The Azure SQL Server needs to be assigned to this group to enable external logins.
 data "azuread_group" "sql_servers" {
-  name = var.sql_servers_group
+  display_name = var.sql_servers_group
 }
 
 # An Azure AD service principal used as Azure Administrator for the Azure SQL Server resource
 resource "azuread_application" "sa" {
-  name     = random_password.sa.keepers.name
-  homepage = "https://test.example.com"
+  display_name = "${local.prefix}-sa"
+  web {
+    homepage_url = "https://test.example.com"
+  }
 }
 
 resource "azuread_service_principal" "sa" {
-  application_id = azuread_application.sa.application_id
+  client_id = azuread_application.sa.client_id
 }
 
 resource "azuread_service_principal_password" "sa" {
-  service_principal_id = azuread_service_principal.sa.id
-  value                = random_password.sa.result
-  end_date_relative    = "360h"
+  service_principal_id = azuread_service_principal.sa.object_id
 }
 
 # An Azure AD service principal used to test creating an external login to the Azure SQL server resource
 resource "azuread_application" "user" {
-  name     = random_password.user.keepers.name
-  homepage = "https://test.example.com"
+  display_name = "${local.prefix}-user"
+  web {
+    homepage_url = "https://test.example.com"
+  }
 }
 
 resource "azuread_service_principal" "user" {
-  application_id = azuread_application.user.application_id
+  client_id = azuread_application.user.client_id
 }
 
 resource "azuread_service_principal_password" "user" {
   service_principal_id = azuread_service_principal.user.id
-  value                = random_password.user.result
-  end_date_relative    = "360h"
 }
 
 # Temporary resource group
@@ -145,11 +129,11 @@ resource "azurerm_mssql_server" "sql_server" {
 
   version                      = "12.0"
   administrator_login          = "SuperAdministrator"
-  administrator_login_password = random_password.sa.result
+  administrator_login_password = azuread_service_principal_password.sa.value
 
   azuread_administrator {
     tenant_id      = var.tenant_id
-    object_id      = azuread_service_principal.sa.application_id
+    object_id      = azuread_service_principal.sa.client_id
     login_username = azuread_service_principal.sa.display_name
   }
 
@@ -163,13 +147,12 @@ resource "azuread_group_member" "sql" {
   member_object_id = azurerm_mssql_server.sql_server.identity[0].principal_id
 }
 
-resource "azurerm_sql_firewall_rule" "sql_server_fw_rule" {
-  count               = length(var.local_ip_addresses)
-  name                = "AllowIP ${count.index}"
-  resource_group_name = azurerm_mssql_server.sql_server.resource_group_name
-  server_name         = azurerm_mssql_server.sql_server.name
-  start_ip_address    = var.local_ip_addresses[count.index]
-  end_ip_address      = var.local_ip_addresses[count.index]
+resource "azurerm_mssql_firewall_rule" "sql_server_fw_rule" {
+  count            = length(var.local_ip_addresses)
+  name             = "AllowIP ${count.index}"
+  server_id        = azurerm_mssql_server.sql_server.id
+  start_ip_address = var.local_ip_addresses[count.index]
+  end_ip_address   = var.local_ip_addresses[count.index]
 }
 
 # The Azure SQL Database used in tests
@@ -230,6 +213,7 @@ output "instance" {
     login_name = mssql_login.server.login_name,
     password   = mssql_login.server.password
   }
+  sensitive = true
 }
 
 
@@ -254,7 +238,7 @@ resource "mssql_user" "database" {
     }
   }
   database = azurerm_mssql_database.db.name
-  username = random_password.database.keepers.username
+  username = "${local.prefix}-user"
   password = random_password.database.result
 }
 
@@ -263,6 +247,7 @@ output "database" {
     username = mssql_user.database.username,
     password = mssql_user.database.password
   }
+  sensitive = true
 }
 
 
@@ -275,7 +260,7 @@ resource "mssql_user" "external" {
     host = azurerm_mssql_server.sql_server.fully_qualified_domain_name
     azure_login {
       tenant_id     = var.tenant_id
-      client_id     = azuread_service_principal.sa.application_id
+      client_id     = azuread_service_principal.sa.client_id
       client_secret = azuread_service_principal_password.sa.value
     }
   }
@@ -286,7 +271,8 @@ resource "mssql_user" "external" {
 output "external" {
   value = {
     tenant_id     = var.tenant_id
-    client_id     = azuread_service_principal.user.application_id
+    client_id     = azuread_service_principal.user.client_id
     client_secret = azuread_service_principal_password.user.value
   }
+  sensitive = true
 }