+++ test-fixtures/all/versions.tf
@@ -1,25 +1,25 @@
 terraform {
-  required_version = "~> 0.13"
+  required_version = "~> 1.5"
   required_providers {
     azuread = {
       source  = "hashicorp/azuread"
-      version = "~> 2.22.0"
+      version = "~> 2.47"
     }
     azurerm = {
       source  = "hashicorp/azurerm"
-      version = "~> 3.8.0"
+      version = "~> 3.85"
     }
     docker = {
       source  = "kreuzwerker/docker"
-      version = "~> 2.16.0"
+      version = "~> 3.0"
     }
     local = {
       source  = "hashicorp/local"
-      version = "~> 2.2"
+      version = "~> 2.4"
     }
     random = {
       source  = "hashicorp/random"
-      version = "~> 3.2"
+      version = "~> 3.6"
     }
   }
 }