+++ test-fixtures/local/versions.tf
@@ -1,13 +1,13 @@
 terraform {
-  required_version = ">= 0.13"
+  required_version = "~> 1.5"
   required_providers {
     docker = {
       source  = "kreuzwerker/docker"
-      version = "~> 2.16.0"
+      version = "~> 3.0"
     }
     local = {
       source  = "hashicorp/local"
-      version = "~> 2.2"
+      version = "~> 2.4"
     }
   }
 }