+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Upgraded to go version 1.21.
+- Upgraded dependencies.
 
 ## [0.2.7] - 2022-12-16