+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Upgraded to go version 1.21.
 - Upgraded dependencies.
+- Replaced github.com/denisenkom/go-mssqldb with github.com/microsoft/go-mssqldb.
 
 ## [0.2.7] - 2022-12-16